# public/api-spec.yaml
openapi: 3.0.0
info:
  title: Report Mapper API
  version: 1.0.0
  description: API for managing wildlife observation reports with geospatial data
  contact:
    name: Report Mapper Team
servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://your-production-domain.com/api
    description: Production server
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: ${KEYCLOAK_ISSUER}/protocol/openid-connect/auth
          tokenUrl: ${KEYCLOAK_ISSUER}/protocol/openid-connect/token
          scopes:
            openid: OpenID Connect scope
            profile: User profile information
            email: User email
      description: OAuth2 authentication via Keycloak
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OAuth2 Bearer token from Keycloak (for direct API access)
    cookieAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: Session cookie (automatically set by browser)
  schemas:
...
    ImageRevision:
      type: object
      properties:
        id:
          type: string
          description: Unique image identifier
        revisionId:
          type: integer
          description: Revision number
        imageKey:
          type: string
          description: Path to object store location of image
        description:
          type: string
          description: Image description
        location:
          type: object
          properties:
            latitude:
              type: number
              format: double
            longitude:
              type: number
              format: double
        metadataCreatedAt:
          type: string
          format: date-time
          description: Date and time from image metadata
        createdAt:
          type: string
          format: date-time
          description: Original image creation date
        updatedAt:
          type: string
          format: date-time
          description: Last update date
        revisionCreatedAt:
          type: string
          format: date-time
          description: Revision creation date
        published:
          type: boolean
        submitted:
          type: boolean
        owner:
          type: string
        canEdit:
          type: boolean
        canDelete:
          type: boolean
        canPublish:
          type: boolean
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
security:
  - oauth2: [openid, profile, email]
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /images/{id}/revisions:
    post:
      summary: Create a new revision of an existing image
      description: |
        Create a new revision from an existing image.

        **Access Control**: Only the owner with edit permission can create a revision.

        **Immutable Properties** (copied from base image, cannot be changed):
        - id (stays the same across all revisions)
        - image_key (cannot change file location)
        - created_at (original creation date)
        - owner (ownership cannot be transferred via revisions)

        **Mutable Properties** (can be changed in the new revision):
        - description
        - location (image_metadata_location)
        - published/submitted status
      tags:
        - Images
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The image ID to create a revision for
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                location:
                  type: object
                  properties:
                    latitude:
                      type: number
                    longitude:
                      type: number
      responses:
        '201':
          description: New revision created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  revisionId:
                    type: integer
                  immutableProperties:
                    type: object
                    properties:
                      imageId:
                        type: string
                      imageKey:
                        type: string
                      createdAt:
                        type: string
                        format: date-time
                      owner:
                        type: string
                  mutableProperties:
                    type: object
                    properties:
                      description:
                        type: string
                      location:
                        type: object
                        properties:
                          latitude:
                            type: number
                          longitude:
                            type: number
                  revisionCreatedAt:
                    type: string
                    format: date-time
                  published:
                    type: boolean
                  submitted:
                    type: boolean
        '400':
          description: Invalid request - attempting to change immutable property
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Authentication required
        '403':
          description: Forbidden - only owner can create revisions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Image not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /images/{id}/{revisionId}:
    get:
      summary: Get a specific image revision
      description: Retrieve a specific revision by image ID and revision ID
      tags:
        - Images
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: revisionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Image revision details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageRevision'
        '403':
          description: Forbidden - insufficient permissions
        '404':
          description: Image revision not found

    put:
      summary: Update a specific image revision
      description: Update metadata of an unpublished image revision
      tags:
        - Images
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: revisionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                location:
                  type: object
                  properties:
                    latitude:
                      type: number
                    longitude:
                      type: number
      responses:
        '200':
          description: Revision updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  revisionId:
                    type: integer
                  description:
                    type: string
                  location:
                    type: object
                  updatedAt:
                    type: string
                    format: date-time
        '400':
          description: Cannot edit published revision
        '401':
          description: Authentication required
        '403':
          description: Forbidden - insufficient permissions
        '404':
          description: Image revision not found

    delete:
      summary: Delete a specific image revision
      description: Delete an unpublished image revision
      tags:
        - Images
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: revisionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Revision deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '400':
          description: Cannot delete published revision
        '401':
          description: Authentication required
        '403':
          description: Forbidden - insufficient permissions
        '404':
          description: Image revision not found

    patch:
      summary: Publish or submit an image revision
      description: |
        Perform an action on a specific image revision:
        - **publish**: Publish the revision (requires publish permission)
        - **submit**: Submit the revision for review (requires ownership)
      tags:
        - Images
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: revisionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - action
              properties:
                action:
                  type: string
                  enum: [publish, submit]
                  description: The action to perform
      responses:
        '200':
          description: Action completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  revisionId:
                    type: integer
                  published:
                    type: boolean
                  submitted:
                    type: boolean
                  message:
                    type: string
        '400':
          description: Invalid action
        '401':
          description: Authentication required
        '403':
          description: Forbidden - insufficient permissions
        '404':
          description: Image revision not found
  /images/{id}/file:
    get:
      summary: Get the actual image file
      description: |
        Retrieve the image file from storage. Returns the binary image data.
        Optionally specify a revision ID to get a specific revision, otherwise returns the published revision.
      tags:
        - Images
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The image ID
        - name: revisionId
          in: query
          required: false
          schema:
            type: integer
          description: Optional revision ID (defaults to published revision)
      responses:
        '200':
          description: Image file
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/gif:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
        '403':
          description: Forbidden - insufficient permissions
        '404':
          description: Image not found
...